feat(SDK): add new headset types and move to base headset sdk
The Headsets enum has been moved from the DeviceFinder to the
SDK BaseHeadset script to keep in line with the controllers enum
being in the SDK BaseController script.

Additional Headset types have been added for GearVR supported devices
and for Daydream and Cardboard. As Google devices don't report
differences between the devices the summary and detail both return
the same response of `Daydream` or `GoogleCardboard`.