feat(Locomotion): ability to move, rotate, scale play area by dragging
The Drag World script mimics the movement mechanic popularised in
Sports Bar VR (Hulk Mode) or in Tilt Brush and Google Blocks, where the
PlayArea can be moved, rotated and scaled based on dragging and
grabbing the world with the controllers and moving them in a certain
way to determine the appropriate mechanism to be performed.

A new custom data type has been created that allows the state of a
Vector3 to contain boolean information, this is used in this case for
knowing which axis on the Vector3 to lock, but could also be used
for other boolean information relating to a Vector3.