fix(Internal): add editor datatypes and attributes to VRTK namespace
To prevent name collisions with 3rd party scripts, the editor datatypes
and attributes used by VRTK are now within the VRTK namespace.

The Serialized Property Extensions script is also now included in the
VRTK namespace in case 3rd party scripts are also using this same
script causing it to be in the project twice.