chore(Interaction): remove physics controllable auto interaction option
The Physics Controllables previously had an `Auto Interaction` option
which would automatically create a Controller Rigidbody Activator on
the control if one was not already present.

This caused a number of issues where the Rigidbody Activator collider
was not of the correct size and would lead to confusion with the
controls. The simplest solution is to just remove the mechanism all
together and expect the Rigidbody Activator prefab to be added
manually when it's required.

Example scene 25 has been updated to reflect this change.