fix(UIPointer): repetitive enter events
When hovering over elements where ExecuteEvents.ExecuteHierarchy does
not return a target, the Enter event would fire repetitively. This
change ensures the event only fires once for these types of elements.
The reason this was happening was because the `pointerEventData` was
not being set properly when firing enter for these elements and would
trigger again the next time Hover was called.

Thanks to @kristianpd for the work on this fix.