feat(Interaction): expose max distance delta parameter on track grab
The Track Object Grab Mechanic had an internal parameter for the
Max Distance Delta used within the MoveTowards method. However, this
was always set to `10f` but on occasion it was prudent to change this
value to something else. The simplest solution is to just expose this
parameter to allow the user to set it to what they need it to be.