fix(Pointers): ensure controller reference set when controller found
The Controller Reference was only set when the activation button was
pressed which meant that the enter event would not emit until the
activation button was pressed. This caused an issue when the
`Activate On Enable` option was set as the Controller Reference
had not been set at that point.