feat(SDK): add option to exclude target build platforms
There was an issue with Unity 2017.2 where it would display a warning
when trying to manage the VR settings due to the Facebook and Switch
platform not being supported.

The new Target Platform Group Exclusions will allow for these two
platforms to be automatically excluded but also the array can be
updated with additional platforms if required.