feat(SDK): add WindowsMR support for SteamVR
The SteamVR SDK also supports Windows Mixed Reality devices but again
has some slight differences such as controller collider shapes and
controller orientations.

The Windows Mixed Reality controller also has an additional thumbstick
as well as a touchpad. The thumbstick is now referred to as
`TouchpadTwo` but only has axis change events as the button press is
reserved as a system event in SteamVR and there is no touch sensing
available.