feat(SDK): add touchpad two touch state
The thumbstick (touchpad two) on the WindowsMR controller doesn't
have a touch sensor but it is prudent to know when it is being
touched so a touchpad two touch button event can be created.

This is useful for enabling the pointers when the thumbstick is being
used as the default input for teleporting on WindowsMR is usually
pressing the thumbstick.

This has been implemented by just checking if the axis on the
thumbstick is over a set threshold. The threshold is embedded in
the SteamVR Controller class so isn't overly easy to change at this
moment.