feat(Structure): add ObsoleteInspector attribute
The new `[ObsoleteInspector]` attribute is placed above public
inspector parameters in place of the previous `[HideInInspector]`
attribute as hidden deprecated parameters was causing confusion
and this new attribute will display the parameter still but render
the label in a red, italic font so it's obvious there is something
different about the field. The tooltip for the parameter is also
replaced by the obsolete message from the `[System.Obsolete]`
attribute.