fix(UI): handle worlds without camera
When working with additive scenes there is null exception if:

- first scene has a UICanvas, but no camera
- second scene is the one with a camera

As soon as the first scene is loaded, the raycast gives an error
because trying to do an update but there is no camera.
The error stops after the second scene is loaded.

I checked the implementation of the unity GraphicRaycaster
and they have a check for the eventCamera been null,
in which case, they use a width/height of the diplay.