feat(Interaction): select objects by distance from attach point
Adds the multiTouchSelectionStyle option to VRTK_InteractTouch
(see #315 and #1618).
Allows user to specify which interactable will be
selected, when the controller is over multiple objects. The
existing style selects the most recently touched object.
This commit adds the ClostestOrigin style, which selects the
object with the closest origin to the controller attach point.