feat(Examples): consolidate example scenes and provide descriptions
The example scenes have been reduced in number to fewer scenes that
better show off the capabilities and provide text to understand what
each section of the scene is showcasing.

A number of hotfixes have been applied to certain scripts as they
were not working correctly if the SDK Manager was additively
loaded in from another scene.

All original example files have been moved to the `LegacyExampleFiles`
directory.