fix(SDKManager): Unity 2018 Requires "None" Device to Be Last
As it is right now "None" will be put at the top of the list in
Player Settings > XR Devices > VR Enabled Devices. In 2017 this
seemed be bypassed by default, but in 2018 the assumed intent is to
not start in VR mode.

This fix puts "None" last when the settings are auto managed.