chore(SDK): provide new location to Oculus SDK download
The Oculus SDK for Unity is now available as a Unity asset on the Unity
Asset Store. This package includes the Oculus Avatar package as well so
out of the box provides controller visualisation support.

The documentation has been updated to point to this Unity asset as the
basis for getting Oculus SDK support rather than using the Oculus
Utilities download from the Oculus developer website.