fix(Examples): display error if script define symbols have not been set
There is an issue where the new example scenes don't set the required
VRTK script define symbols because this is done via the SDK Manager
script and that is no longer present in the example scenes as it is
dynamically loaded in via the constructor scene.

This fix simply throws an error if the example scene cannot find the
required scripting define symbols and attempts to provide useful
information for the user to rectify the issue.