fix(Examples): remove script define symbol check and provide popup
The previous way of attempting to notify the user that they need to
open the example constructor scene before opening any of the example
scenes did not work if no 3rd party SDK was installed that required
it. It would simply just continually throw the error message.

The new solution is not to try and second guess what the set up is
automatically and to just provide a popup window with information
for the user on requiring them to open the constructor example scene.