chore(structure): create base structure for v4.0.0
The new structure for the repo removes all Unity project settings
and any Visual Studio project settings and just includes VRTK
example scenes and relevant documentation.

The `VRTK.Unity.Core` repo is now a submodule that includes all of
the core VRTK scripts and additional files.

Everything else has been deleted and will be added back in when
necessary.

A TestRunner scene has been added that can be used to run the
Unit Tests if required as the TestRunner scene is a completely
empty scene.