fix(ArtificialControls): No longer resets value on re-enable
A SetValue function has been introduced to the base class for all controls to implement  to define how their values should be set (ex: position, rotation).
This has been implemented for both the ArtificialSlider and ArtificialRotator but not the ArtificialPusher as this object is meant to reset to the original value after a press anyway.