feat(Examples): add locomotion and climbing example scene
The Locomotion example scene showcases body representation which
provides a physical representation of the player's body in virtual
space and can be used to prevent walking through walls and determine
collisions and when the player becomes diverged from their virtual
representation (e.g. by walking in real life through a virtual wall).

The example scene also contains a demo of the climbing prefab along
with a basic version of an axis control to move the play area around
with the touchpad/thumbstick.