feat(Examples): material swap highlighting on interactions
The example scenes have been updated to show how to swap a material
on a MeshRenderer during an interaction such as touching or grabbing.

The Interactions example scene (004) has been updated with the
Flying Saucer prefab now highlights the different meshes in different
colours when it is touched.

The Locomotion example scene (005) has been updated with the Climbing
Hanger prefab now highlights when touched and highlights a different
colour when it is grabbed.