fix(physics): change collision detection mode for 2018.3
Unity 2018.3 seems to have a bug where it does not support continuous
dynamic for the rigidbody collision detection, instead it uses a newly
introduced mode called collision speculative.

This fix wraps all of the places where continuous dynamic is used with
ifdefs to check if the version is Unity 2018.3 or higher and if it is
then it will use the continuous speculative mode instead.