fix(Interactions): apply rotational velocity from existing drive spin
The rotational velocity applied to the rotational drive when it is
released is now no longer taken from the velocity of the interactor
but instead the current velocity the drive has is carried over into
continuing the drive rotation at that speed.

The issue with using the controller velocities is that the controller
velocity isn't always the correct way when applied to the direction
of the rotation.

For example, grabbing a z axis rotation on the left and pulling up to
apply upward velocity applies a positive rotation, but grabbing the
same z axis rotation object on the right and pulling up still applies
that same velocity which is still a positive rotation, causing the
rotation to go in the wrong way.