refactor(Pointers): remove pointer prefab materials directory
The Materials directory for the Pointers prefab has now been removed
and the materials have been relocated to the Prefabs/SharedResources
directory as these materials can (and are) used for objects other than
the pointers.

The included shaders have also been tidied up to all be in a VRTK
folder along with consistent formatting and the ability to prevent
Unity from auto updating the shader code.