feat(Helpers): colliders that track source location within physics
The TrackedCollider prefab allows a collider to follow a source
GameObject in the scene using rigidbody physics so it adheres to the
physics rules within the world.

This is useful for setting up controller colliders that cannot directly
follow the controller 1:1 if the controller goes through geometry.

This allows a better interaction with joint objects.