feat(examples): add warehouse door with interactable buttons to control
A warehouse door has been added that can be slid up and down based on
pressing physics buttons. This showcases the TrackedCollider prefab as
well as the controller colliders are only enabled when in proximity of
the buttons and will use rigidbody physics to track the position of
the controller. This means that the controller colliders cannot pass
through the buttons and therefore won't prevent the unity physics from
glitching.