feat(Prefabs): add SnapDropZone prefab
The SnapDropZone prefab allows "snapping" an object to a set
transform target state and specifically comes with support for
interactable object so they can be snapped and unsnapped by
ungrabbing and grabbing respectively. Rigidbodies on the snapped
objects will be set to kinematic und unset when unsnapping
automatically. A simple checkbox allows turning on cloning of the
unsnapped object to make the SnapDropZone act as some kind of
instantiation component.