perf(general): prevent repeated access of Unity property
Many of Unity's types offer properties that return data from the
native part of the engine. Accessing such a property requires a
transition of that data from native code and as such does a managed
heap allocation and potentially an expensive lookup or computation
operation. To reduce this overhead the property values are cached
into local variables through this change.