fix(general): flip usage of FloatRange to reflect Zinnia changes
The FloatRange component has changed so the default constructor is
(0f, 0f) and therefore no longer requires the `FloatRange.Zero` static
helper, so any occurrence of `FloatRange.Zero` is simply not required.
However, this means that any previous usage of the old constructor
which expected a (`Float.MinValue`, `Float.MaxValue`)
also needs to be updated to not use the default constructor but instead
to use the new `FloatRange.MinMax` static helper.