docs(README.md): add guidance of how to run example scene
The README doc has been updated with a guide of how to run the farmyard
example scene along with informing users to run the scene in `Maximize
On Play` mode to reduce the performance issues observed running in the
Unity Editor.

A few style changes have been made as well to improve markdown
consistency.