docs(structure): apply style changes to standardize markdown
The style of the markdown has been made consistent throughout the
various documents as there were a number of inconsistencies like
bullet points being preceded with spaces (of varying length) and
note fields as well. It's better to just have all elements start
with no preceding spaces.

Sentences all end with full stops now as well to make it clear they
have an end to them.

Any use of the word Unity has been updated to conform with the Unity
brand guidelines.