docs(README.md): improve cloning the repo instructions
The instructions for how to clone the repo missed out a step that
requires switching to the cloned repo directory before initializing
the submodule. Whilst this may seem obvious, many people who do not
have experience of git will follow the instructions to the letter
and will have no knowledge of why it's not working.

A Documentation section has been added that clearly links to the
VRTK Academy as well for extra clarity.