fix(Interactions): prevent exception when attempting auto grab
The auto Grab method on the interactor and interactable was throwing
an exception due to the attempt creation of an ActiveCollisionContainer
was failing as the event data list was empty yet it was trying to set
the first element on a list.

This fix simply checks to see if the list is empty and if it is then
adds a default entry that can be set by the creation logic.