feat(Locomotion): add destination location prefab
The DestinationLocation prefab is a way of predefining a visual volume
in the scene that can be interacted with to provide visual feedback
about where a teleportation location is.

The Destination Location is usually used in conjunction with
ObjectPointers which interact with the Destination Locations in the
scene via the DestinationLocationManager.