fix(examples): enable use float values for simulator
The SimulatedCameraRig did not emit float actions for the Use action
so the guns would not fire. This has been fixed by simply converting
the button Boolean value to a float.

The Grab/Use simulator buttons have also been swapped so Grab is
Right Mouse button and Use is Left Mouse Button. This feels more
natural as firing a gun with the LMB is expected as this is usually
the control scheme found in games.