fix(examples): ensure pointer obscurance disables teleporter
The logic to disable teleporting with pointer obscurance was not
working due to some changes in how the teleport action had been
set up. The fix is to nest the pointer select actions in another
container GameObject and it is then this container that is disabled
by the obscurance query.