fix(Interactions): treat interactable as only valid active collision
The Interactor ActiveCollisionsContainer now has a new Validity Rule
that only allows interactions to occur with anything that is a type
of Interactable. This means an Interactor can collide with any other
valid collider in the scene but it can only actively interact with
items that pass the new Validity Rule.