fix(Interactions): optionally inherit RB kinematic on follow action
The GrabInteractableFollowAction can set the kinematic state of the
consumer rigidbody when the follow action is inactive. This is the
default kinematic state for the rigidbody. Initially this default state
was always whatever the actual consumer rigidbody kinematic state was
initially set to, however there may be times where this inherited
default state may not be required so an additional option has been
added to the GrabInteractableFollowAction to ignore the setting of
the inactive state to that of the consumer rigidbody kinematic state.

This fix is largely based on the work by @JWNJWN but also contains
a fix from @bruno1308 which ensures the `ConfigureFollowTracking`
method is called when the GrabSetup is changed.

Co-authored-by: JWNJWN <jwneicho@gmail.com>