fix(examples): set tools to force non kinematic on release from grab
When the tools (saw, hacksaw and hammer) are dropped they should always
assume to be non-kinematic. However, there is an issue where when they
are snapped to the Cloning machine they are kinematic and the
"Will Inherit Is Kinematic When Inactive From Consumer Rigidbody" is
true, then when the cloned object is grabbed it will inherit this
kinematic state as being the default inactive state, so the tools
can't be dropped and have gravity applied when they are released.

A number of fixes could be supplied, like having the cloning machine
force this, but this is just an example and the easiest solution is
to just prevent them from inheriting their kinematic state from the
rigidbody.