ci: implement continuous delivery
Since VRTK is now a UPM package it should automatically be released
as one. This change adds automatic creation of UPM packages for
VRTK, including automatic SemVer-styled versioning based on the
commit messages. A release is both uploaded to the ExtendReality npm
GitHub registry as well as an GitHub release (archived .zip). The
`package.json` file has been updated to publish beta releases out of
the `master` branch, instead of using the CD template's defaults.