feat(examples): add pointer distant grabbers
The pointer distance grabbers can be activated by double clicking the
grip button, which now turns the interactor avatar blue to denote the
pointer distance grabber is enabled. Then from a distance of 5m
the pointer can select a distant interactable and grab it to the
interactor pointing at it.