feat(examples): add bow and arrow to the scene
A new bow and arrow has been added to the scene where the bow has
limited custom scripting but has an interactable string and uses a
snap zone as a nock position for an interactable arrow.

The arrow also has limited scripting just to control the flight of
the arrow.